pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kompile"
version = "0.1.0"
description = "Personal knowledge compiler — turns AI conversations into a structured, cited wiki"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
kompile = "kompile.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["kompile*"]
